requirements.txt
yt-dlp
webvtt-py
openai
streamlit

